custom-dev: 部分需改造的PR中，一些有用的代码先复制过来。

+++ config/seata-config-core/src/main/java/io/seata/config/Configuration.java
@@ -33,7 +33,7 @@ public interface Configuration {
      * @param timeoutMills the timeout mills
      * @return the short
      */
-    short getShort(String dataId, int defaultValue, long timeoutMills);
+    short getShort(String dataId, short defaultValue, long timeoutMills);
 
     /**
      * Gets short.
@@ -207,7 +207,6 @@ public interface Configuration {
     boolean putConfig(String dataId, String content, long timeoutMills);
 
     /**
-     *
      * @param dataId       the data id
      * @param defaultValue the default value
      * @param timeoutMills the timeout mills
@@ -293,5 +292,4 @@ public interface Configuration {
     default String getConfigFromSysPro(String dataId) {
         return System.getProperty(dataId);
     }
-
 }
